VIVO-1049 Move the themes from the webapp to the home directory.
Modify build.xml to put the themes in the home directory instead of the webapp.
Modify ThemeInfoSetup so the ThemeInfo will be based on the home directory.
Modify FreemarkerConfiguration and FreemarkerTemplateLoader to look in the home directory for templates.
Modify I18n and VitroResourcebundle to look for language files in the home directory.
Modify web.xml and add ThemeInfoSetup to intercept requests to vivo/themes, and serve the files from the home directory.